+++ spec/hatchet/rubies_spec.rb
@@ -1,45 +1,9 @@
 require_relative '../spec_helper'
 
 describe "Ruby Versions on cedar-14" do
-  it "should allow patchlevels" do
-    app = Hatchet::Runner.new('mri_193_p547', stack: "cedar-14")
-    app.deploy do |app|
-      version = '1.9.3p547'
-      expect(app.output).to match("ruby-1.9.3-p547")
-      expect(app.run('ruby -v')).to match(version)
-    end
-  end
-
-  it "should deploy ruby 1.9.2 properly" do
-    app = Hatchet::Runner.new('mri_192', stack: "cedar-14")
-    app.deploy do |app|
-      version = '1.9.2'
-      expect(app.output).to match(version)
-      expect(app.run('ruby -v')).to match(version)
-    end
-  end
-
-  it "should deploy ruby 1.9.3 properly" do
-    app = Hatchet::Runner.new('mri_193', stack: "cedar-14")
-    app.deploy do |app|
-      version = '1.9.3'
-      expect(app.output).to match(version)
-      expect(app.run('ruby -v')).to match(version)
-    end
-  end
-
-  it "should deploy ruby 2.0.0 properly" do
-    app = Hatchet::Runner.new('mri_200', stack: "cedar-14")
-    app.deploy do |app|
-      version = '2.0.0'
-      expect(app.output).to match(version)
-      expect(app.run('ruby -v')).to match(version)
-
-      expect(app.output).to match("devcenter.heroku.com/articles/ruby-default-web-server")
-    end
-  end
-
   it "should deploy jruby 1.7.16.1 (jdk 7) properly on cedar-14 with sys props file" do
+    pending("Port this to a more recent stack")
+
     app = Hatchet::Runner.new("ruby_193_jruby_17161_jdk7", stack: "cedar-14")
     app.deploy do |app|
       expect(app.output).to match("Installing JVM: openjdk-7")
@@ -49,19 +13,45 @@ describe "Ruby Versions on cedar-14" do
 end
 
 describe "Ruby versions" do
-  it "should deploy jdk 8 on heroku-18  by default" do
-    app = Hatchet::Runner.new("ruby_193_jruby_1_7_27", stack: "heroku-18")
-    app.deploy do |app|
-      expect(app.output).to match("Installing JVM: openjdk-8")
-      expect(app.output).to match("JRUBY_OPTS is:  -Xcompile.invokedynamic=false")
-      expect(app.output).not_to include("OpenJDK 64-Bit Server VM warning")
+  it "should deploy jdk on heroku-20" do
+    Hatchet::Runner.new("default_ruby", stack: "heroku-20").tap do |app|
+      app.before_deploy do |app|
+        Pathname("Gemfile.lock").write(<<~EOM)
+         GEM
+           remote: https://rubygems.org/
+           specs:
+             rack (2.2.2)
+             rake (13.0.1)
 
-      run!('git commit -am "redeploy" --allow-empty')
-      app.set_config("JRUBY_BUILD_OPTS" => "--dev")
-      app.push!
-      expect(app.output).to match("JRUBY_OPTS is:  --dev")
+         PLATFORMS
+           java
 
-      expect(app.run("ls vendor/jvm/jre/lib/ext")).to match("pgconfig.jar")
+         DEPENDENCIES
+           rack
+           rake
+
+
+         RUBY VERSION
+            ruby 2.5.7p0 (jruby 9.2.13.0)
+        EOM
+
+        Pathname("Rakefile").write(<<~'EOM')
+          task "assets:precompile" do
+            puts "JRUBY_OPTS is: #{ENV['JRUBY_OPTS']}"
+          end
+        EOM
+      end
+
+      app.deploy do
+        expect(app.output).to match("JRUBY_OPTS is: -Xcompile.invokedynamic=false")
+
+        app.set_config("JRUBY_BUILD_OPTS" => "--dev")
+        app.commit!
+        app.push!
+        expect(app.output).to match("JRUBY_OPTS is: --dev")
+
+        expect(app.run("ls .jdk/jre/lib/ext/")).to match("pgconfig.jar")
+      end
     end
   end