+++ spec/hatchet/stack_spec.rb
@@ -3,8 +3,8 @@ require_relative "../spec_helper"
 describe "Stack Changes" do
   xit "should reinstall gems on stack change" do
     Hatchet::Runner.new('default_ruby', stack: "heroku-18").deploy do |app|
-      app.update_stack("heroku-16")
-      run!('git commit --allow-empty -m "heroku-16 migrate"')
+      app.update_stack("heroku-20")
+      run!('git commit --allow-empty -m "heroku-20 migrate"')
 
       app.push!
 
@@ -14,8 +14,8 @@ describe "Stack Changes" do
   end
 
   it "should not reinstall gems if the stack did not change" do
-    Hatchet::Runner.new('default_ruby', stack: "heroku-16").deploy do |app|
-      app.update_stack("heroku-16")
+    Hatchet::Runner.new('default_ruby', stack: "heroku-20").deploy do |app|
+      app.update_stack("heroku-20")
       run!(%Q{git commit --allow-empty -m "cedar migrate"})
 
       app.push!